Add files via upload
Ciel Consulting is developing a new online ordering app. The app will enable users to place orders on any of
the products from Ciel Consulting using any Internet-enabled device, e.g., a laptop, a tablet, or a mobile
phone. This app will work on all operating systems around the world and will require at least a 2G network.
You are appointed as the Business Analyst for the project. Your first task is to identify all the various
requirement categories that this app has to meet and how.
Ciel Consulting has asked you to develop a plan for the order collection software it is going to establish. Use
the template provided to complete a brief plan.