Add files via upload
City to Country (CTC) Realty is a growing realty company that focuses on business properties outside of
major metropolitan areas. It specializes in matching up the buyers and sellers of hobby farms, lake resorts,
bed-and-breakfast businesses, and other types of small businesses located in or around mid-to-small sized
towns. CTC Realty has been in business for several years and has achieved significant growth while interest
rates have been low.
However, it knows that low interest rates will not remain unchanged, and it wants to leverage technology
to compete against larger commercial and residential realty companies. Customers CTC Realty has two
types of customers spread throughout North America: buyers and sellers. Sellers include those who are
often planning to downsize after raising a family or retiring and want something smaller and easier to
manage. Buyers could be any person or business.
Buyers and sellers pay for the services utilized, such as contractual assistance, realtor assistance, etc. and
then pay the appropriate fee to the lenders. CTC receives 1 percent of the total selling price of the property
at closing. The business goals, objectives, and requirements of CTC Realty are to grow substantially to be
able to compete with larger national firms. It wants to increase its market share by 15 percent in the
coming year and decrease the duration of its sale and purchase process from an average of 20 days down
to 10 days.
CTC Realty has determined that a self-service website will help meet these goals and objectives by
providing a communication vehicle for buyers and sellers of real estate to meet and share information. The
application will support all aspects of a real estate sale and also supply advice and recommendations. This
includes connecting lenders, legal analysts, and financial analysts with buyers and sellers.
Various parties will not need to meet face to face but will be connected virtually through the app. For
brevity, it must list only the selling and security processes. In the selling process, the website will offer the
ability to a seller to complete the entire property sales process. When a person is ready to sell a property,
he/she will be able to go online to make the necessary arrangements.
The seller may choose to select a realtor to help through the sales process, represent themselves, and have
legal counsel create the appropriate documents, or they may use the standard document templates found
on the website. The seller can list the property online by selecting the desired level of service and entering
the property specifications and information. The system will then provide a final confirmation number and
make the property available for buyers to peruse. The seller will then have to wait for an offer.
The system will notify both the seller and the seller’s realtor (if applicable) of any offers. At this time, the
seller or the seller’s realtor may submit a counter offer. Once the offer is accepted, the process moves to
closing, which could be a little complex. All final documentation is created for the closing. The buyer and
the seller (and all the selected support staff) review the contracts and agree to the sale. The buyer and the
seller request the system to assign PINs to electronically approve all closing documents. The buyer
and the seller then approve the closing contracts by entering their pin numbers. The system records the completion of the sale and registers the property in the buyer’s name. For security, each user of the system will have a password and a user ID. The user will use their email address as their user ID and create a unique 6–8 digit password. The system will also provide electronic pin numbers to the buyer, the seller, and the appropriate realtors to electronically sign the e-documents.