Add files via upload
Q1. You are working as a Business Analyst in a company that provides online food ordering services for the
customers. Further, the company is a start-up and in the past few months it has been getting a lot of users
as well as a tremendous growth. So, the company wants to you to make a structure for their business
model in order to present it in front of various stakeholders and investors to attract them from investing in
their company.
1. As a Business Analyst your responsibility is to gather all the requirements for your company that
would help them to make effective strategy for accomplishing the task that can be presented in
the form of diagrams.
2. Create a UML diagram using Microsoft Visio in order to represent the business model.